+++ src/parsing/RetailLogHandler.ts
@@ -55,9 +55,6 @@ export default class RetailLogHandler extends LogHandler {
       .on('ARENA_MATCH_START', async (line: LogLine) => {
         await this.handleArenaStartLine(line);
       })
-      // .on('ARENA_MATCH_END', async (line: LogLine) => {
-      //   await this.handleArenaEndLine(line);
-      // })
       .on('CHALLENGE_MODE_START', async (line: LogLine) => {
         await this.handleChallengeModeStartLine(line);
       })
@@ -70,14 +67,14 @@ export default class RetailLogHandler extends LogHandler {
       .on('SPELL_CAST_SUCCESS', async (line: LogLine) => {
         this.handleSpellCastSuccess(line);
       })
-      .on('WAL_ARENA_END', (event: IArenaMatch) => {
-        this.handleWALArenaEnd(event);
+      .on('WAL_ARENA_END', async (event: IArenaMatch) => {
+        await this.handleWALArenaEnd(event);
       })
-      .on('WAL_SHUFFLE_END', (event: IShuffleMatch) => {
-        this.handleWALShuffleMatchEnd(event);
+      .on('WAL_SHUFFLE_END', async (event: IShuffleMatch) => {
+        await this.handleWALShuffleMatchEnd(event);
       })
-      .on('WAL_MALFORMED', (event: IMalformedCombatData) => {
-        this.handleWALParserMalformed(event);
+      .on('WAL_MALFORMED', async (event: IMalformedCombatData) => {
+        await this.handleWALParserMalformed(event);
       });
   }
 
@@ -499,16 +496,17 @@ export default class RetailLogHandler extends LogHandler {
 
   private async handleWALArenaEnd(event: IArenaMatch) {
     console.info('[RetailLogHandler] Got IArenaMatch event from WAL');
-    console.info(event);
+
+    // The last line from the WAL event is the ARENA_MATCH_END line.
     const endLine = new LogLine(event.rawLines[event.rawLines.length - 1]);
     await this.handleArenaEndLine(endLine);
   }
 
   private async handleWALShuffleMatchEnd(event: IShuffleMatch) {
     console.info('[RetailLogHandler] Got IShuffleMatch event from WAL');
-    console.info(event);
-    const lastRound = event.rounds[5];
 
+    // The last line from the WAL event is the ARENA_MATCH_END line.
+    const lastRound = event.rounds[5];
     const endLine = new LogLine(
       lastRound.rawLines[lastRound.rawLines.length - 1]
     );